a few things (read description)
- proper support for legacy shockwave and crimson pots
- buffed revamped shockwave and made it more consistent
- monoliths spawn their pillars
- block *every* raincoat zombie from spawning
- some supporting code for attraidies

+++ Projectiles/Shockwave.cs
@@ -16,13 +16,14 @@ namespace tsorcRevamp.Projectiles {
             projectile.tileCollide = false;
             projectile.damage = 60;
             projectile.friendly = true;
-            projectile.penetrate = 3;
+            projectile.penetrate = 99;
             projectile.alpha = 255;
-            projectile.usesIDStaticNPCImmunity = true;
-            projectile.idStaticNPCHitCooldown = 20; // a shockwave projectile can't damage an enemy more frequently than once every 20 ticks to prevent "shotgunning" by landing directly on an enemy
+            projectile.usesIDStaticNPCImmunity = true; // only one of these can hit a target at once
+            projectile.idStaticNPCHitCooldown = 20; // after which it becomes immune to this projectile for 20 frames
         }
 
         public override void AI() {
+            projectile.ai[0]++;
             int dust = Dust.NewDust(new Vector2(projectile.position.X, projectile.position.Y - 10), projectile.width, projectile.height, 31, 0, 0, 100, default, 1.0f);
             Main.dust[dust].noGravity = true;
             int dust2 = Dust.NewDust(new Vector2(projectile.position.X, projectile.position.Y - 20), projectile.width, projectile.height, 31, 0, 0, 100, default, 1.0f);
@@ -30,9 +31,11 @@ namespace tsorcRevamp.Projectiles {
             int dust3 = Dust.NewDust(new Vector2(projectile.position.X, projectile.position.Y - 30), projectile.width, projectile.height, 31, 0, 0, 100, default, 1.0f);
             Main.dust[dust3].noGravity = true;
 
-            projectile.damage = (int)(projectile.damage * 0.925f); //scale down the damage as it gets further from the impact location
-            if (projectile.damage <= 2) {
-                projectile.Kill();
+            if (projectile.ai[0] > 10) { //delay the damage reduction (feels better)
+                projectile.damage = (int)(projectile.damage * 0.925f); //scale down the damage as it gets further from the impact location
+                if (projectile.damage <= 2) {
+                    projectile.Kill();
+                } 
             }
         }
     }
